video增加分页，部分修改

+++ app-video/src/main/java/com/jdyx/app/appvideo/controller/PostController.java
@@ -17,7 +17,6 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import redis.clients.jedis.JedisPool;
 
 import java.util.List;
-import java.util.Map;
 
 @Slf4j
 @RequestMapping("/post")
@@ -53,8 +52,10 @@ public class PostController {
             @ApiImplicitParam(paramType="query", name = "name", value = "岗位名称", required = true, dataType = "String"),
     })
     public Object savePost(Post post){
+        if (post.getName()==null){
+            return ResultUtil.exceptionMap("2019","岗位名称无效");
+        }
         try {
-//            Post post = new Post(null, name);
             postService.savePost(post);
             log.info("添加岗位的id ：{}",post.getId());
             return ResultUtil.successMap("");
@@ -71,6 +72,9 @@ public class PostController {
             @ApiImplicitParam(paramType="query", name = "id", value = "岗位id", required = true, dataType = "String"),
     })
     public Object  deletePost(Integer id){
+        if (id==null){
+            return ResultUtil.exceptionMap("2019","岗位id无效");
+        }
         try {
             postService.deletePost(id);
             return ResultUtil.successMap("");
